requirements.txt
aiosmtplib==2.0.2
annotated-types==0.6.0
anyio==3.7.1
astroid==3.0.1
bcrypt==4.0.1
blinker==1.7.0
blis==0.7.11
boto3==1.34.7
botocore==1.34.7
catalogue==2.0.10
Cerberus==1.3.5
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cloudpathlib==0.16.0
colorama==0.4.6
confection==0.1.4
cryptography==41.0.7
cymem==2.0.8
Deprecated==1.2.14
dill==0.3.7
distlib==0.3.7
dnspython==2.4.2
docopt==0.6.2
ecdsa==0.18.0
email-validator==2.1.0.post1
fastapi==0.104.1
fastapi-mail==1.4.1
greenlet==3.0.2
h11==0.14.0
httpcore==1.0.2
httptools==0.6.1
httpx==0.25.2
idna==3.6
importlib-resources==6.1.1
iniconfig==2.0.0
isort==5.13.0
Jinja2==3.1.2
jmespath==1.0.1
langcodes==3.3.0
limits==3.7.0
MarkupSafe==2.1.3
mccabe==0.7.0
murmurhash==1.0.10
numpy==1.26.2
packaging==23.2
passlib==1.7.4
pep517==0.13.1
pip-api==0.0.30
pipreqs==0.4.13
platformdirs==4.1.0
plette==0.4.4
pluggy==1.3.0
preshed==3.0.9
psycopg2==2.9.9
pyasn1==0.5.1
pycparser==2.21
pydantic==2.5.2
pydantic-settings==2.1.0
pydantic_core==2.14.5
pytest==7.4.3
python-dateutil==2.8.2
python-dotenv==1.0.0
python-jose==3.3.0
python-multipart==0.0.6
PyYAML==6.0.1
requests==2.31.0
requirementslib==3.0.0
rsa==4.9
ruff==0.1.9
s3transfer==0.10.0
setuptools==69.0.2
six==1.16.0
slowapi==0.1.8
smart-open==6.4.0
sniffio==1.3.0
spacy==3.7.2
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SQLAlchemy==2.0.23
sqlmodel==0.0.14
srsly==2.4.8
starlette==0.27.0
thinc==8.2.2
time-machine==2.13.0
tomlkit==0.12.3
tqdm==4.66.1
typer==0.9.0
typing_extensions==4.9.0
tzdata==2023.3
urllib3==2.0.7
uvicorn==0.24.0.post1
wasabi==1.1.2
watchfiles==0.21.0
weasel==0.3.4
websockets==12.0
wrapt==1.16.0
yarg==0.1.9
